scaling_policy/requirements.txt
google-cloud-compute 
crewai